+++ src/g_input.h
@@ -31,7 +31,7 @@ void G_ClearCarry(void);
 extern short (*G_CarryAngleTic)(double angle);
 extern short (*G_CarryAngle)(double angle);
 void G_UpdateAngleFunctions(void);
-int G_CarryPitch(double pitch);
+short G_CarryPitch(double pitch);
 int G_CarrySide(double side);
 int G_CarryVert(double vert);
